NP-3585 Move project row to own component
